Add tailored resume for Microsoft About the job OPG Next is a multidisciplinary team doing strategic thinking and exploration to help shape how AI will be used in the workplace of the future. You are a software engineer early in your career and passionate about building the next generation of software experiences can touch millions of people.
We are looking for a Software Engineer to join our team. You will be part of a dynamic and agile group dedicated to innovation by learning, exploring and building working code. In this role you have the opportunity to contribute to high-impact innovations. This may be transforming concepts to reality by building prototypes or helping to take our ideas to production. You will have the chance to tackle challenges up and down the tech stack. We achieve our goals as a team and embrace motivated independent thinkers that are comfortable operating amidst ambiguity and a rapidly changing technical landscape. You are a strong problem solver with a stronger desire to learn.

We are a small team with a culture of support and collaboration and a can do attitude. We embrace a growth mindset and encourage the team and leaders to bring their best each day. Join us and help shape the future of the world.

Microsoft